# Fix performance bugs and add a benchmarking stub