--- a/llvm-external-projects/iree-dialects/include/iree-dialects/Transforms/TransformMatchers.h
+++ b/llvm-external-projects/iree-dialects/include/iree-dialects/Transforms/TransformMatchers.h
@@ -569,8 +569,10 @@ private:
 //===---------------------------------------------------------------------===//
 
 struct MatchedReductionCaptures {
-  int64_t rank;
-  int64_t reductionDimensionSize;
+  int64_t reductionRank = 0;
+  int64_t reductionDimensionSize = 0;
+  int64_t maybeLeadingRank = 0;
+  int64_t maybeTrailingRank = 0;
 };
 
 /// Creates a group of matchers for: