This revision introduces a 3 stage stategy for mapping reductions
along with a simple heuristic that finds the right tile sizes and vector sizes as a function
of the reduction size, the max number of threads and the warp shuffle size.
    
This provides a first set of tunable knobs for this strategy that can evolve into something more general in the future.
    
A simple benchmarking stub file is added that can be used with `sed`-based commands.
It is currently unclear where the control for such benchmarking should live and what relevant sizes should be.

Example of usage using the current IREE pipeline:
```
cat tests/transform_dialect/cuda/benchmark_linalg_reductions.stub.mlir | \
sed "s/\${SZ1}/1024/g" | \
sed "s/\${SZ2}/1024/g" | \
iree-compile - --iree-hal-target-backends=cuda --iree-hal-benchmark-dispatch-repeat-count=5 | \
nvprof --print-gpu-trace iree-run-module --entry_function=reduction_2d_static --device=cuda --function_input="1024x1024xf32=1" 2>&1 | \
grep reduction
```
reports (note the super slow first run that pulls data from CPU memory):
```
EXEC @reduction_2d_static
==426762== Profiling application: iree-run-module --entry_function=reduction_2d_static --device=cuda --function_input=1024x1024xf32=1
423.07ms  3.1230ms           (1024 1 1)       (256 1 1)        16        0B       32B  NVIDIA GeForce          1        13                     -                -  reduction_2d_static_dispatch_0_generic_1024x1024 [33]
426.19ms  7.2960us           (1024 1 1)       (256 1 1)        16        0B       32B  NVIDIA GeForce          1        13                     -                -  reduction_2d_static_dispatch_0_generic_1024x1024 [34]
426.20ms  6.8810us           (1024 1 1)       (256 1 1)        16        0B       32B  NVIDIA GeForce          1        13                     -                -  reduction_2d_static_dispatch_0_generic_1024x1024 [35]
426.21ms  7.0080us           (1024 1 1)       (256 1 1)        16        0B       32B  NVIDIA GeForce          1        13                     -                -  reduction_2d_static_dispatch_0_generic_1024x1024 [36]
426.21ms  7.0090us           (1024 1 1)       (256 1 1)        16        0B       32B  NVIDIA GeForce          1        13                     -                -  reduction_2d_static_dispatch_0_generic_1024x1024 [37]
```

Example of usage with this PR:
```
cat tests/transform_dialect/cuda/benchmark_linalg_reductions.stub.mlir | \
sed "s/\${SZ1}/1024/g" | \
sed "s/\${SZ2}/1024/g" | \
iree-compile - --iree-hal-target-backends=cuda --iree-codegen-llvmgpu-enable-transform-dialect-jit --iree-hal-benchmark-dispatch-repeat-count=5 | \
nvprof --print-gpu-trace iree-run-module --entry_function=reduction_2d_static --device=cuda --function_input="1024x1024xf32=1" 2>&1 | \
grep reduction
```
reports (note the super slow first run that pulls data from CPU memory):
```
EXEC @reduction_2d_static
==426762== Profiling application: iree-run-module --entry_function=reduction_2d_static --device=cuda --function_input=1024x1024xf32=1
423.72ms  3.3241ms           (1024 1 1)       (256 1 1)        16        0B  1.2500KB  NVIDIA GeForce          1        13                     -                -  reduction_2d_static_dispatch_0_generic_1024x1024 [33]
427.05ms  7.5850us           (1024 1 1)       (256 1 1)        16        0B  1.2500KB  NVIDIA GeForce          1        13                     -                -  reduction_2d_static_dispatch_0_generic_1024x1024 [34]
427.05ms  7.3280us           (1024 1 1)       (256 1 1)        16        0B  1.2500KB  NVIDIA GeForce          1        13                     -                -  reduction_2d_static_dispatch_0_generic_1024x1024 [35]
427.06ms  7.1690us           (1024 1 1)       (256 1 1)        16        0B  1.2500KB  NVIDIA GeForce          1        13                     -                -  reduction_2d_static_dispatch_0_generic_1024x1024 [36]
427.07ms  7.2000us           (1024 1 1)       (256 1 1)        16        0B  1.2500KB  NVIDIA GeForce          1        13                     -                -  reduction_2d_static_dispatch_0_generic_1024x1024 [37]
```